+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     allow stripping out these unsupported tags
 - **MP4**: Renamed `AdvisoryRating::None` to `AdvisoryRating::Inoffensive`
 - Renamed `TaggedFile::remove_tag` to `TaggedFile::take`
+- **Vorbis Comments**: `VorbisComments::insert_picture` now accepts a user provided `PictureInformation`
 
 ### Fixed
 - **MP4**: Non-full `meta` atoms are now properly handled.